Move base model and QS into project package to avoid circular imports
